Moved mixins into a separate directory
