+++ src/python_pachyderm/mixin/__init__.py
@@ -1,0 +1,13 @@
+"""
+Exposes a mixin for each pachyderm service. These mixins should not be used
+directly; instead, you should use `python_pachyderm.Client`. The mixins exist
+exclusively in order to provide better code organization (because we have
+several mixins, rather than one giant `Client` class.)
+"""
+
+from python_pachyderm.mixin.pfs import PFSMixin
+from python_pachyderm.mixin.pps import PPSMixin
+from python_pachyderm.mixin.transaction import TransactionMixin
+from python_pachyderm.mixin.version import VersionMixin
+from python_pachyderm.mixin.admin import AdminMixin
+from python_pachyderm.mixin.util import Service, GRPC_MODULES, PROTO_MODULES
