+++ src/python_pachyderm/mixin/admin.py
@@ -1,6 +1,6 @@
 from python_pachyderm.proto.admin import admin_pb2_grpc as admin_grpc
 from python_pachyderm.proto.pps import pps_pb2 as pps_proto
-from python_pachyderm.util import Service
+from .util import Service
 
 
 class AdminMixin:
