+++ src/python_pachyderm/mixin/pfs.py
@@ -3,8 +3,7 @@ import itertools
 from contextlib import contextmanager
 
 from python_pachyderm.proto.pfs import pfs_pb2 as pfs_proto
-from python_pachyderm.util import commit_from
-from python_pachyderm.util import Service
+from .util import Service, commit_from
 
 
 BUFFER_SIZE = 20 * 1024 * 1024
